Move foo,bar
